requirements.txt
Django==2.2
psycopg2==2.8.2
django-money==0.14.4
djangorestframework==3.9.2
django-filter==2.1.0
